added: gaussian_blur_parallel implementation
In short, we split the buffer containing the image data based on however
many threads (`JOBS`) we want to create.

By dividing the image into independent width ranges, each thread
processes a different part of the buffer, so there is no need for mutual
exclusion.

The variable names are still weird, but will fix them later.